refactor: streamline imports and enhance formula logic
- Removed unnecessary import statements from various files to clean up the codebase.
- Introduced a new helper module for vectorized logical operations in taxation formulas, improving readability and maintainability.
- Updated existing formulas to utilize the new helper functions, ensuring consistent logic across the codebase.
- Added a script for creating or updating pull requests automatically, enhancing the development workflow.
