+++ openfisca_france_entreprises/variables/caracteristiques_unite_legale.py
@@ -5,7 +5,6 @@ A variable is a property of an Entity such as a Etablissement, a UniteLegale…
 See https://openfisca.org/doc/key-concepts/variables.html
 """
 
-# Import from openfisca-core the Python objects used to code the legislation in OpenFisca
 from openfisca_core.model_api import (
     MONTH,
     YEAR,
@@ -14,7 +13,6 @@ from openfisca_core.model_api import (
     set_input_dispatch_by_period,
 )
 
-# Import the Entities specifically defined for this tax and benefit system
 from openfisca_france_entreprises.entities import UniteLegale
