+++ openfisca_france_entreprises/variables/taxes/taxation_energies/taxation_autres_produits_energetiques.py
@@ -5,40 +5,15 @@ A variable is a property of an Entity such as a Etablissement, a UniteLegale…
 See https://openfisca.org/doc/key-concepts/variables.html
 """
 
-# Import from numpy the operations you need to apply on OpenFisca's population vectors
-# Import from openfisca-core the Python objects used to code the legislation in OpenFisca
-from functools import reduce
+from openfisca_core.model_api import YEAR, Variable, select
 
-from numpy import logical_and, logical_or
-from openfisca_core.model_api import YEAR, Variable, not_, select
-
-# Import the Entities specifically defined for this tax and benefit system
 from openfisca_france_entreprises.entities import Etablissement
-
-
-def _and(*args):
-    r = args[0]
-    for a in args[1:]:
-        r = logical_and(r, a)
-    return r
-
-
-def _or(*args):
-    r = args[0]
-    for a in args[1:]:
-        r = logical_or(r, a)
-    return r
-
-
-def _not(x):
-    return not_(x)
-
-
-def _dep_in(departement, codes):
-    """Vectorized: True where departement is in codes."""
-    if len(codes) == 1:
-        return departement == codes[0]
-    return reduce(lambda a, b: a | b, (departement == c for c in codes))
+from openfisca_france_entreprises.variables.taxes.formula_helpers import (
+    _and,
+    _dep_in,
+    _not,
+    _or,
+)
 
 
 class taxe_interieure_consommation_sur_produits_energetiques(Variable):
