+++ openfisca_france_entreprises/variables/taxes/taxation_energies/tccfe/taux_2022.py
@@ -2,7 +2,6 @@
 
 from openfisca_core.model_api import YEAR, Variable
 
-# Import the Entities specifically defined for this tax and benefit system
 from openfisca_france_entreprises.entities import Etablissement
