+++ openfisca_france_entreprises/variables/taxes/taxation_energies/tdcfe/taux_2020.py
@@ -1,9 +1,7 @@
 """Variables and formulas for this module."""
 
-# Import from openfisca-core the Python objects used to code the legislation in OpenFisca
 from openfisca_core.model_api import YEAR, Variable
 
-# Import the Entities specifically defined for this tax and benefit system
 from openfisca_france_entreprises.entities import Etablissement
 
 # NB : le code départementale de Mayotte change de 143 en 2014 à 106 en 2015
